requirements.txt
google-genai
python-dotenv
groq
